リファクタリング
- commandをcomand.rsに分離
- menuをmenu.rsに分離
- 選択したファイルのパスのチェックをDocument::set_file_pathに移行
